removed nltk tokenizer
It was removed as it was added due to the slow speed of the spacy tokenizer. However, it turns out that the spacy tokenizer is fast enough (more than enough in fact), but the serialization of the doc object is slow.
Thus once should paralize the filter instead of tokenization.